refactor: js -> jsx
