+++ frontend/sucpi/src/Admin/setting/QSetting.jsx
