+++ frontend/sucpi/src/Student/global-challenge/GlobalChallenge.jsx
