+++ frontend/sucpi/src/components/FilterTable/FilterTable.jsx
