+++ frontend/sucpi/src/components/Graph/BarGraph/BarChart.jsx
