+++ frontend/sucpi/src/components/Graph/PieGraph/AfterGraph.jsx
