+++ frontend/sucpi/src/components/Graph/PieGraph/BeforeGraph.jsx
