+++ frontend/sucpi/src/components/Graph/PieGraph/ThreeQChart.jsx
