+++ frontend/sucpi/src/components/Sidebar/Sidebar.jsx
@@ -1,4 +1,3 @@
-// Sidebar.jsx
 import React from 'react';
 import { Link, useLocation } from 'react-router-dom';
 import './Sidebar.css'; // CSS 파일 임포트
